requirements.txt
Django==4.2.15
django-crispy-forms==2.3
pytz==2024.2
requests==2.32.3
sweetify==2.3.1

django-filter~=25.1